create_movie_method_works, but a unique rating has only one foreign movie

+++ MovieApiMvc/Models/DomainModels/Rating.cs
@@ -2,13 +2,13 @@ namespace MovieApiMvc.Models.DomainModels;
 public class Rating
 {
     public Guid Id { get; private set; }
-    public double? Kp { get;private set; }
-    public double? Imdb { get;private set; }
-    public double? FilmCritics { get;private set; }
-    public double? RussianFilmCritics { get;private set; }
+    public short? Kp { get;private set; }
+    public short? Imdb { get;private set; }
+    public short? FilmCritics { get;private set; }
+    public short? RussianFilmCritics { get;private set; }
     public Guid MovieId { get; private set; }
     public Movie? Movie { get; private set; }
-    private Rating(Guid id, double? kp, double? imdb, double? filmCritics, double? russianFilmCritics, Guid movieId, Movie? movie)
+    private Rating(Guid id, short? kp, short? imdb, short? filmCritics, short? russianFilmCritics, Guid movieId, Movie? movie)
     {
         Id = id;
         Kp = kp;
@@ -17,7 +17,7 @@ public class Rating
         RussianFilmCritics = russianFilmCritics;
     }
 
-    public static Rating Create(Guid movieId, double? kp = 0, double? imdb = 0, double? filmCritics = 0, double? russianFilmCritics = 0, Movie? movie = null)
+    public static Rating Create(Guid movieId, short? kp = 0, short? imdb = 0, short? filmCritics = 0, short? russianFilmCritics = 0, Movie? movie = null)
     {
         if (kp is null && imdb is null && filmCritics is null && russianFilmCritics is null)
         {
